protection.md: Cryptography Capabilities and Support section
protection.md: Edited for consistent use of leading statement throughout technical capabilities document. Split apart some of the compound capabilities statements. Reworded to clarify other statements. Since this is a catalog of capabilities, I removed the "Requires some or all of the following depending on implementation:" statements throughout. Removed "for the IOT device" throughout to be consistently worded with other capabilities categories.